utils: Move the lower verbose name check to tests
